refactor: move proxy tests into internal/proxy as in-package tests
git mv + package clause change. In-package (package proxy) tests keep
unexported access; no test logic changed.
