+++ internal/proxy/auth_gokrb5_test.go
@@ -1,4 +1,4 @@
-package main
+package proxy
 
 import (
 	"os"
