+++ internal/proxy/auth_gss_darwin_test.go
@@ -1,6 +1,6 @@
 //go:build darwin
 
-package main
+package proxy
 
 import (
 	"testing"
