+++ internal/proxy/framing_conformance_test.go
@@ -1,4 +1,4 @@
-package main
+package proxy
 
 import (
 	"bytes"
