+++ internal/proxy/gokrb5_integration_test.go
@@ -1,4 +1,4 @@
-package main
+package proxy
 
 import (
 	"bufio"
