+++ internal/proxy/hop_by_hop_test.go
@@ -1,4 +1,4 @@
-package main
+package proxy
 
 import (
 	"bufio"
