+++ internal/proxy/response_fuzz_test.go
@@ -1,4 +1,4 @@
-package main
+package proxy
 
 import (
 	"bufio"
