+++ internal/proxy/upstream_tls_test.go
@@ -1,4 +1,4 @@
-package main
+package proxy
 
 import (
 	"bufio"
